feat(programs): implement PythCoreAdapter and add PythLazer component placeholder
